// frontend/src/client/models/ImageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 图片响应模型
 */
export type ImageResponse = {
    id: string;
    image_type: string;
    url: string;
    script_id: number;
    author_id: number;
    positive_prompt: (string | null);
    negative_prompt: (string | null);
    width: (string | null);
    height: (string | null);
    created_at: (string | null);
};